Fix stop_time importer
Importing stop-times.txt was only using the trip_id for the uID, this caused the script to create only a single stop time record for the first stop_time it found for a trip_id. This fix corrects this by adding 'stop_id' as a second uID to sto_times.txt in the fileMap. The script now correctly creates a record for each stoptime in stop_time.txt.